FE/Refactor import paths in LandingHeader and page.tsx
