+++ app/(pages)/page.tsx
@@ -1,9 +1,9 @@
 import Image from "next/image";
 import { Mail, ChevronRight, User2 as User } from "lucide-react";
-import { Hero } from "../components/ui/icons";
+import { Hero } from "../client/components/landing/icons";
 import { Button } from "../client/components/ui/button";
-import { LandingHeader } from "../client/components/LandingHeader";
-import { Footer } from "../client/components/footer";
+import { LandingHeader } from "../client/components/landing/LandingHeader";
+import { Footer } from "../client/components/landing/Footer";
 
 export default function Landing() {
   return (
